Move getCustomProps into core
