+++ packages/popover/src/js/helpers/index.js
@@ -1,5 +1,4 @@
 export * from "./applyPositionStyle";
-export * from "./getCustomProps";
 export * from "./getDelay";
 export * from "./getMiddlewareOptions";
 export * from "./getPadding";
